Make BlendContext inner values to private,
Move FilterContext::concat to BlendContext->concat_into
